Updated
Staging the current progress

From now on, try to implement a simple minimax algorithm with a evaluation based on number of empty hexagon cells only

The main focus is to make sure that minimax recursion is working.

And the team need to determine either use tree structure or recursion.

For now, prefer implemented via recursion.